+++ tests/feature_flag_management_tests.rs
@@ -15,8 +15,11 @@
 mod test_helpers;
 use crate::test_helpers::*;
 use std::error::Error;
+
 #[test]
 fn test_enable_a_feature_flag() -> Result<(), Box<dyn Error>> {
+    skip_if_rabbitmq_version_below!(3, 13, 0);
+
     let ff_name = "detailed_queues_endpoint";
 
     run_succeeds(["feature_flags", "enable", "--name", ff_name]);
@@ -27,6 +30,8 @@ fn test_enable_a_feature_flag() -> Result<(), Box<dyn Error>> {
 
 #[test]
 fn test_enable_all_stable_feature_flags() -> Result<(), Box<dyn Error>> {
+    skip_if_rabbitmq_version_below!(4, 0, 0);
+
     let ff_name = "rabbitmq_4.0.0";
 
     run_succeeds(["feature_flags", "enable_all"]);