+++ tests/memory_breakdown_tests.rs
@@ -18,6 +18,8 @@ use crate::test_helpers::*;
 
 #[test]
 fn test_show_memory_breakdown_in_bytes_succeeds() -> Result<(), Box<dyn Error>> {
+    skip_if_rabbitmq_version_below!(4, 0, 0);
+
     let rc = api_client();
     let nodes = rc.list_nodes()?;
     let first = nodes.first().unwrap();
@@ -40,6 +42,8 @@ fn test_show_memory_breakdown_in_bytes_succeeds() -> Result<(), Box<dyn Error>> 
 
 #[test]
 fn test_show_memory_breakdown_in_percent_succeeds() -> Result<(), Box<dyn Error>> {
+    skip_if_rabbitmq_version_below!(4, 0, 0);
+
     let rc = api_client();
     let nodes = rc.list_nodes()?;
     let first = nodes.first().unwrap();